requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pillow==10.4.0
torch==2.4.1
torchvision==0.19.1
numpy==2.1.2
scikit-learn==1.5.2
pandas==2.2.3
tqdm==4.66.5

